Moved tooltip modifier to core accessibility utils
